+++ CHANGELOG.md
@@ -19,6 +19,7 @@ https://github.com/nwnxee/unified/compare/build8193.21...HEAD
 - Effect: GetTrueEffect()
 - Effect: RemoveEffectById()
 - Effect: ReplaceEffectByIndex()
+- Player: SetTlkOverride()
 
 ### Changed
 - The argument stack is now cleared after every NWNX function call.